#2513 move logic to Atom Tool
