+++ packages/ketcher-react/src/script/ui/state/handleHotkeysOverItem.ts
@@ -6,10 +6,7 @@ import {
   fromBondsAttrs,
   fromFragmentDeletion,
   FunctionalGroup,
-  Atom,
-  Action,
-  SGroup,
-  fromSgroupDeletion
+  Atom
 } from 'ketcher-core'
 import { STRUCT_TYPE } from 'src/constants'
 import { openDialog } from './modal'
@@ -19,8 +16,6 @@ import { Editor } from '../../editor'
 import { updateSelectedAtoms } from 'src/script/ui/state/modal/atoms'
 import { fromAtom, toAtom, fromBond, toBond } from '../data/convert/structconv'
 import SGroupTool from '../../editor/tool/sgroup'
-import { getGroupIdsFromItemArrays } from '../../editor/tool/helper/getGroupIdsFromItems'
-import { Selection } from '../../editor/Editor'
 
 type TNewAction = {
   tool?: string
@@ -40,12 +35,6 @@ type HandleHotkeyOverItemProps = {
   newAction: TNewAction
   editor: Editor
   dispatch: Dispatch
-}
-
-type HandleHotkeyOverSGroupsProps = {
-  selected: Selection
-  atomProps: Partial<TNewAction>
-  editor: Editor
 }
 
 export function handleHotkeyOverItem(props: HandleHotkeyOverItemProps) {
@@ -272,49 +261,6 @@ function handleBondTool({ hoveredItemId, newAction, editor }: HandlersProps) {
     { label: 'C' }
   )[0]
   editor.update(newBond)
-}
-
-export function handleHotkeyOverSGroup({
-  selected,
-  atomProps,
-  editor
-}: HandleHotkeyOverSGroupsProps) {
-  const struct = editor.render.ctab
-  const selectedSGroupsId =
-    selected && getGroupIdsFromItemArrays(struct.molecule, selected)
-  const action = new Action()
-  const atomsInSGroups: number[] = []
-  const sgroups = struct.sgroups
-  const functionalGroups = struct.molecule.functionalGroups
-
-  selectedSGroupsId.forEach((sGroupId) => {
-    const sGroupItem = sgroups.get(sGroupId)?.item
-    if (
-      FunctionalGroup.isContractedFunctionalGroup(sGroupId, functionalGroups)
-    ) {
-      const atomsWithoutAttachmentPoint =
-        SGroup.getAtomsSGroupWithoutAttachmentPoint(sGroupItem, struct.molecule)
-      atomsInSGroups.push(...atomsWithoutAttachmentPoint)
-      action.mergeWith(fromSgroupDeletion(struct, sGroupId))
-      action.mergeWith(
-        fromFragmentDeletion(struct, {
-          atoms: atomsWithoutAttachmentPoint,
-          bonds: SGroup.getBonds(struct, sGroupItem)
-        })
-      )
-    }
-  })
-  action.mergeWith(
-    fromAtomsAttrs(
-      struct,
-      selected?.atoms?.filter(
-        (selectAtomId) => !atomsInSGroups.includes(selectAtomId)
-      ),
-      atomProps,
-      true
-    )
-  )
-  editor.update(action)
 }
 
 function handleChargeTool({ hoveredItemId, newAction, editor }: HandlersProps) {
